// src/app/shared/lean-learn-api/model/mCQuestion.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MCQuestion { 
    id?: number;
    question?: string;
    possibleAnswers?: Array<string>;
    correctAnswerIndex?: number;
    explanation?: string;
}