// src/app/shared/lean-learn-api/model/mCQuizDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MCQuestionDTO } from './mCQuestionDTO';


export interface MCQuizDTO { 
    questions?: Array<MCQuestionDTO>;
}